# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "API for Real Estate, tool to manage real estate properties and users.",
        "title": "Real Estate API",
        "termsOfService": "http://swagger.io/terms/",
        "contact": {
            "name": "API Support",
            "url": "http://www.swagger.io/support",
            "email": "support@swagger.io"
        },
        "license": {
            "name": "Apache 2.0",
            "url": "http://www.apache.org/licenses/LICENSE-2.0.html"
        }
    },
    "basePath": "/v1",
    "paths": {
...
        "/admin/complaints": {
            "get": {
                "security": [
                    {
                        "ApiKeyAuth": []
                    }
                ],
                "description": "Returns a paginated list of complaints, optionally filtered by status",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "admin"
                ],
                "summary": "List complaints",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Filter by status: new, in_progress, closed",
                        "name": "status",
                        "in": "query"
                    },
                    {
                        "type": "integer",
                        "description": "Limit",
                        "name": "limit",
                        "in": "query"
                    },
                    {
                        "type": "integer",
                        "description": "Offset",
                        "name": "offset",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/github_com_sikozonpc_social_internal_store.Complaint"
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {}
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {}
                    }
                }
            }
        },
        "/admin/complaints/{complaintID}": {
            "get": {
                "security": [
                    {
                        "ApiKeyAuth": []
                    }
                ],
                "description": "Returns a single complaint with target name and author info",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "admin"
                ],
                "summary": "Get complaint by ID",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "Complaint ID",
                        "name": "complaintID",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/github_com_sikozonpc_social_internal_store.Complaint"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {}
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {}
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {}
                    }
                }
            }
        },
        "/admin/complaints/{complaintID}/status": {
            "patch": {
                "security": [
                    {
                        "ApiKeyAuth": []
                    }
                ],
                "description": "Admin changes the status of a complaint",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "admin"
                ],
                "summary": "Update complaint status",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "Complaint ID",
                        "name": "complaintID",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "New status",
                        "name": "payload",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/cmd_api.UpdateComplaintStatusPayload"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/github_com_sikozonpc_social_internal_store.Complaint"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {}
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {}
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {}
                    }
                }
            }
        },
...
        "/complaints": {
            "post": {
                "security": [
                    {
                        "ApiKeyAuth": []
                    }
                ],
                "description": "User submits a complaint about a listing or company",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "complaints"
                ],
                "summary": "Create a complaint",
                "parameters": [
                    {
                        "description": "Complaint data",
                        "name": "payload",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/cmd_api.CreateComplaintPayload"
                        }
                    }
                ],
                "responses": {
                    "201": {
                        "description": "Created",
                        "schema": {
                            "$ref": "#/definitions/github_com_sikozonpc_social_internal_store.Complaint"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {}
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {}
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {}
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {}
                    }
                }
            }
        },
...
    },
    "definitions": {
...
        "cmd_api.CreateComplaintPayload": {
            "type": "object",
            "required": [
                "description",
                "target_id",
                "target_type",
                "type"
            ],
            "properties": {
                "description": {
                    "type": "string",
                    "maxLength": 2000
                },
                "target_id": {
                    "type": "integer"
                },
                "target_type": {
                    "type": "string",
                    "enum": [
                        "listing",
                        "company"
                    ]
                },
                "type": {
                    "type": "string",
                    "enum": [
                        "incorrect_listing",
                        "rule_violation",
                        "fraud"
                    ]
                }
            }
        },
...
        "cmd_api.UpdateComplaintStatusPayload": {
            "type": "object",
            "required": [
                "status"
            ],
            "properties": {
                "status": {
                    "type": "string",
                    "enum": [
                        "new",
                        "in_progress",
                        "closed"
                    ]
                }
            }
        },
...
        "github_com_sikozonpc_social_internal_store.Complaint": {
            "type": "object",
            "properties": {
                "author_name": {
                    "description": "joined from users.username",
                    "type": "string"
                },
                "created_at": {
                    "type": "string"
                },
                "description": {
                    "type": "string"
                },
                "id": {
                    "type": "integer"
                },
                "status": {
                    "description": "\"new\", \"in_progress\", \"closed\"",
                    "type": "string"
                },
                "target_id": {
                    "type": "integer"
                },
                "target_name": {
                    "description": "joined from listings.title or companies.name",
                    "type": "string"
                },
                "target_type": {
                    "description": "\"listing\" | \"company\"",
                    "type": "string"
                },
                "type": {
                    "description": "\"incorrect_listing\", \"rule_violation\", \"fraud\"",
                    "type": "string"
                },
                "updated_at": {
                    "type": "string"
                },
                "user_id": {
                    "type": "integer"
                }
            }
        },
...
    },
    "securityDefinitions": {
        "ApiKeyAuth": {
            "type": "apiKey",
            "name": "Authorization",
            "in": "header"
        }
    }
}